+++ packages/events/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @happyvertical/smrt-events
 
+## 0.38.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.38.19
+  - @happyvertical/smrt-assets@0.38.19
+  - @happyvertical/smrt-places@0.38.19
+  - @happyvertical/smrt-profiles@0.38.19
+  - @happyvertical/smrt-tenancy@0.38.19
+  - @happyvertical/smrt-types@0.38.19
+  - @happyvertical/smrt-ui@0.38.19
+
 ## 0.38.18
 
 ### Patch Changes