+++ packages/personas/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @happyvertical/smrt-personas
 
+## 0.38.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.38.19
+  - @happyvertical/smrt-agents@0.38.19
+  - @happyvertical/smrt-tenancy@0.38.19
+
 ## 0.38.18
 
 ### Patch Changes