+++ packages/secrets/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @happyvertical/smrt-secrets
 
+## 0.38.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.38.19
+  - @happyvertical/smrt-tenancy@0.38.19
+
 ## 0.38.18
 
 ### Patch Changes