+++ packages/smrt-ios/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @happyvertical/smrt-ios
 
+## 0.38.19
+
 ## 0.38.18
 
 ## 0.38.17